Move integration tests to separate subdirectory
